requirements.txt
Flask
gunicorn
pytesseract
opencv-python
Pillow